requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pyguitarpro==0.10.0
midiutil==1.2.1
